SSCS-10286 move messaging models to model package
