+++ src/test/java/uk/gov/hmcts/reform/sscs/service/ccdupdate/HearingsJourneyServiceTest.java
@@ -8,8 +8,8 @@ import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
 import uk.gov.hmcts.reform.sscs.exception.GetCaseException;
 import uk.gov.hmcts.reform.sscs.exception.UpdateCaseException;
-import uk.gov.hmcts.reform.sscs.messaging.HearingUpdate;
-import uk.gov.hmcts.reform.sscs.messaging.HmcMessage;
+import uk.gov.hmcts.reform.sscs.model.messaging.HearingUpdate;
+import uk.gov.hmcts.reform.sscs.model.messaging.HmcMessage;
 import uk.gov.hmcts.reform.sscs.model.single.hearing.HearingGetResponse;
 import uk.gov.hmcts.reform.sscs.service.CcdCaseService;
 import uk.gov.hmcts.reform.sscs.service.HmcHearingService;
